requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
install==1.3.5
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.12.2
